up bai tap

+++ src/ss12/exercises/product/model/SortDownProduct.java
@@ -5,6 +5,6 @@ import java.util.Comparator;
 public class SortDownProduct implements Comparator<Product> {
     @Override
     public int compare(Product o1, Product o2) {
-        return (int) (o2.getPrice() - o1.getPrice());
+        return (o2.getPrice() - o1.getPrice());
     }
 }
